修改物流配送逻辑

+++ sugo-takeout-bean/src/main/java/com/sugo/takeout/bean/dto/SellerAdditionalDto.java
@@ -34,5 +34,5 @@ public class SellerAdditionalDto implements Serializable {
     /**
      * 配送时间
      */
-    private Long avgDeliveryTime;
+    private Double avgDeliveryTime;
 }
